requirements.txt
Django>=1.9.4
djangorestframework
markdown
django-filter
django-rest-swagger
tweepy
pyenchant
